move miner invariants check
